use both onfocus and onclick for input field (+tests for preconnect)
I originally went with a click handler instead of a focus handler
as that way we naturally handle both "focus after click" and "click
when already focussed" (noting that focus doesn't fire onclick when
already focussed), which is useful as a way to re-open a closed
listbox.

Switch this to a regular focus handler (with the latter preserved
in a separate click handler) so that "focus after slash key" also
enjoys the preconnect optimisation.

== Size ==

* Parent commit, base:    2323 kB
* This change, essence:   2330 kB (+7 B)
* This change, committed: 2319 kB (-11 B)