Fix bug where title-only match in grouped mode has an empty link label
In grouped mode, lvl0 is used for the category heading, and lvl1-5
make up the link label. However, this assumed at least one of lvl1-5
matched, which is not technically required.

In practice, this is still a poor quality result though as ideally
we'd render the (first) lvl1 on the page even if it isn't the part
that matched the query. But, that requires a change in scraper
configuration. When operating a site and using minibar with grouped
mode, you should configure the scraper to have:

```
"min_indexed_level": 1,
```

That way, it won't create matchable records with just a lvl0, but
instead it'll create records with at least a lvl0 and lvl1, and then
return the first of those.

== Transfer size ==

Before:
> $ npm run size
> JS  2151

After:
> JS  2153 (+2 B; +0.1%)