improve alignment of icons in search input field
The `--tsmb-size-sm` and `--tsmb-size-base` variable are close
to each other in value by default (1rem and 0.8rem). This made it
difficult to notice that some aspects used the wrong variable for
their alignment.

`.demo-custom-form` now uses a more extreme value (0.5rem) to make
it easier to notice misalignment or other odd-feeling proportions.

The input field is placed to the left by sm + base + sm. This intended
the magnifying glass icon to be a square with "size-base by size-base"
dimensions, and size-sm amount of spacing on either side. Fix
the `left` offset from `--tsmb-size-base` to `--tsmb-size-sm`.

The slash icon was all sorts of weird, including use of bare "em" in
the middle of it, and an oddball "negative 1em" deduction somewhere
that got it roughly in the right place, but I don't recall what that
would've been based on or why that was a good idea. It is a bit of a
complicated case indeed since the intuitive thing would be to make it
"--tsmb-size-input" tall, and use "size-sm" as spacing around, but
the resulting icon would be too small design-wise. It seems in this
case, I think the icon looks better if it stays roughly where it was
but implementing using a more robust calculation, one that doesn't
fall apart when you change the variable values to something entirely
different.

I went with the slash icon being "--tsmb-size-input" plus half of
size-sm for its width and height, and then, naturally, offset by
a quarter of size-sm from the top. In other words, it eats about half
of the size-sm padding that would normally between the slash icon and
the edge of the search input.