set explicit `.tsmb-form` padding
I try to keep opinionated styles on the outer `<form>` element to a
minimum to ease integration. For the form element, we actually
internally depend on this having no padding as it messes with the
layout calculations otherwise.

There is no padding on this element by default in browsers, but when
integrating it into the jquery.com WordPress theme, I noticed that
a "reset" design started interferring with it in a way that seemed
unreasonable and beyond what I expect from a drop-in widget.