set explicit text-decocation on `.tsmb-foot` anchor link
We set this explicitly do this on the anchor links in the search
results (both initial and :hover). We didn't yet do this on the
footer, which meant that on mostly unstyled pages, or sites that
apply text-decoration:underline to anchor links pre-hover, would
leak into the widget, and this would then oddly go away on-hover,
which is kind of the opposite of what you'd expect.

The odd selector rule in the codebase was biased by our demo, so
I've adjusted the demo to be a bit more hostile by default (and
arguably improve accessibility for other links in the demo) to make
this less likely to regress in the future.