CSS: Fix icon overlapping long search input values
Reserve a bit of right padding equal so that typed text and input
cursor don't get obscured behind the "slash" or "close" icons.

When JS is disabled, or when the form is closed and the "slash"
option is not enabled, don't reserve this padding to avoid an
unexplained gap.

It is possible for the "slash" icon and "close" icon to have different
sizes (one is based on "sm" size, the other based on "input" size,
although currently their padding evens out). Right now, this is
setting the padding only based on the "close" icon since that is the
only icon that can appear while the text field is focussed.

I think it feels more smooth if the cut off does not change or shift
when the input is focussed/de-focussed, so I choose not to have
two different paddings.

To avoid an ugly state where old text in the input field is rendered
behind the transparent "slash" icon, I added a solid background color
to the slash icon equal to the input background color, which makes
sure that the slash icon maintains an "on top" appearance.

== npm run size ==

A few unrelated spaces are changed to improve file compression.

CSS:
* Before: 1834
* After:  1869 (+35 B, +2%)