build: fixes for deprecations in GitHub CI
GitHub CI the basic git clone step, not as a built-in or stable feature
but as versioned consumer software.

Further more, the `actions/checkout` software wraps its `git` commands
in a Node.js script. This, of course, couldn't be a standalone Node.js
script, no, we need to also involve more than 80 npm packages as
dependencies. (These are non-dev dependencies, the code in question
has another 1,461 dev dependencies.).

Lastly, we couldn't possibly run this on the base image which GitHub
pre-installs Node.js LTS on. No we must have our own container with
its own copy of Node.js. That way, we can tell the world that
"millions of developers" rely on Docker and Node.js to power their CI.

Our mission? To tell those same millions of non-paying customers
(i.e. "products") that they have dared to use an outdated version of
Node.js and that they must upgrade their CI or else. How? Because of
a chain reaction in incompetence:
* The internal GitHub system that builds the Docker containers
  for `actions/checkout` and other actions, goes through a pointless
  semver-major cycle every six months. Rather than something
  reasonable like https://github.com/nvm-sh/nvm which is able to
  install any arbitrary Node.js version, GitHub's Node.js installer
  is effectively tied to a single Node.js version, so whenever Node.js
  has a major release, GitHub's system goes through a semver-major
  cycle as well, and we must of course deprecate the old thing.
* GitHub then naturally has to update `actions/checkout` to
  internally use the newer Node.js version from time to time.
  This change isn't observable to end-users since the embedded Node.js
  isn't exposed. But, since following semver.org is no longer "cool"
  in 2024, GitHub will do a pointless semver-major bump because
  one of its dependencies did, and the line must go up!
* GitHub then has to "deprecate" the Node.js provisioner to no longer
  have to "support" installing the older Node.js version. Again,
  for a reasonable system like nvm-sh there is nothing to "support".

The end result, you as the end-user simply declaring your CI job
to perform `actions/checkout@v3` will be told that "you" are using
Node.js 16 and must upgrade. With a link to a blog post like this
one:

https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/

It's almost like Microsoft is intentionally creating unusable
software for their competitors to use, to then sell you "AI" to get
halfway back to normal levels of productivity.