Fix inconsistent data attribute location for web component
For the most part 1.3.0, expected data attributes on the `<form>`
element, but this didn't work right for styling where we largely
swap selectors from `.tsmb-form` to `typesense-minibar`. This means
the following selector didn't work right in web component context:

```
.tsmb-form[data-group="true"] .tsmb-suggestion_title

typesense-minibar[data-group="true"] .tsmb-suggestion_title
```

Thinking about it some more, it seems more elegant to have these
on the custom element rather than on the form. So rather than fix
the bug by changing the styles, let's change the place where we
set the data attributes to match the style.

Also use this time to promote the web component by default in
documentation.