remove unused `.tsmb-form--slash::after` style
Follows-up ca5519d369705429ee, which moved this from applying last,
in a media query, to hide the keyboard short cut on narrow viewports
(likely touchscreen devices); to instead apply before the actual style
unconditionally, thus never actually applying as it is immediately
overriden by another style right below it.

The default for `::after`, is already to not exist / be hidden.

Ref https://github.com/jquery/typesense-minibar/pull/2.